[feat] add function prototype (17) - calculate total click from event table - print out favorite topic for every users

+++ demo/src/main/java/com/example/demo/repository/UserClickRepository.java
@@ -6,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserClickRepository extends JpaRepository<UserClick, Long> {
+public interface UserClickRepository extends JpaRepository<UserClick, Integer> {
+
 }
